requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
chromadb>=0.5.5
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
sentence-transformers>=3.0.1
rank-bm25>=0.2.2
